Make more things internal
