+++ packages/common/src/client/sync/options.ts
@@ -48,26 +48,6 @@ export interface SyncOptions {
   crudUploadThrottleMs?: number;
 }
 
-/**
- * @internal
- */
-export type ResolvedSyncOptions = Required<SyncOptions>;
-
-/**
- * @internal
- */
-export function resolveSyncOptions(options: SyncOptions): ResolvedSyncOptions {
-  return {
-    appMetadata: options.appMetadata ?? {},
-    connectionMethod: options.connectionMethod ?? SyncStreamConnectionMethod.HTTP,
-    fetchStrategy: options.fetchStrategy ?? FetchStrategy.Buffered,
-    params: options.params ?? {},
-    includeDefaultStreams: options.includeDefaultStreams ?? true,
-    retryDelayMs: options.retryDelayMs ?? 5000,
-    crudUploadThrottleMs: options.crudUploadThrottleMs ?? 1000
-  };
-}
-
 // TODO: This should not be part of SyncOptions. Remove the WebSocket options from @powersync/common into a separate
 // package and make this an option only available when creating a custom remote.
