+++ packages/shared-internals/src/db/ConnectionClosedError.ts
