+++ packages/tanstack-react-query/package.json
@@ -33,6 +33,7 @@
   "homepage": "https://docs.powersync.com",
   "peerDependencies": {
     "@powersync/common": "workspace:^1.54.0",
+    "@powersync/shared-internals": "workspace:^1.0.0",
     "react": "*"
   },
   "dependencies": {
