+++ packages/tanstack-react-query/tsconfig.json
@@ -11,6 +11,9 @@
   "references": [
     {
       "path": "../common"
+    },
+    {
+      "path": "../shared-internals"
     }
   ],
   "include": ["src/**/*"]
