+++ packages/vue/tsconfig.json
@@ -14,6 +14,9 @@
     },
     {
       "path": "../web"
+    },
+    {
+      "path": "../shared-internals"
     }
   ],
   "include": ["src/**/*"]
