+++ packages/web/src/worker/sync/WorkerClient.ts
@@ -1,4 +1,3 @@
-import { ResolvedSyncOptions } from '@powersync/common';
 import * as Comlink from 'comlink';
 import { getNavigatorLocks } from '../../shared/navigator.js';
 import {
@@ -8,7 +7,7 @@ import {
   SharedSyncInitOptions,
   WrappedSyncPort
 } from './SharedSyncImplementation.js';
-import { SubscribedStream } from '@powersync/shared-internals';
+import { SubscribedStream, ResolvedSyncOptions } from '@powersync/shared-internals';
 
 /**
  * A client to the shared sync worker.
